Переносит функцию отправки данных в модуль работы с сетью.
