+++ js/main.js
@@ -1,7 +1,7 @@
 import './map.js';
-import './form.js';
 import './restrictions.js';
 import './server-data.js';
+import './form.js';
 import './generation-data.js';
-import './test.js';
+import './filters.js';
