+++ js/server-data.js
@@ -1,10 +1,11 @@
 //TODO: Модуль для обмена данными по сети.
-import {getFilter} from './test.js';
+import {getFilter} from './filters.js';
 import {showAlert} from './utils.js';
-import {formEnableHandler} from './form.js';
-
+import {formEnableHandler, sendForm, showMessage} from './form.js';
+import {restoreParameters} from './map.js';
 
 const DATA_RECEVE_ADDRESS = 'https://23.javascript.pages.academy/keksobooking/data';
+const FORM_SEND_ADDRESS = 'https://23.javascript.pages.academy/keksobook';
 
 const createFetch = (onError = showAlert) => {
   fetch(DATA_RECEVE_ADDRESS, {
@@ -19,3 +20,27 @@ const createFetch = (onError = showAlert) => {
 
 createFetch();
 
+sendForm.addEventListener('submit', (evt) => {
+  evt.preventDefault();
+  const errorMessage = document.querySelector('#error');
+  const successMessage = document.querySelector('#success');
+  const errorMessageShow = errorMessage.cloneNode(true);
+  const successMessageShow = successMessage.cloneNode(true);
+  const formData = new FormData(evt.target);
+  fetch(
+    FORM_SEND_ADDRESS,
+    {
+      method: 'POST',
+      body: formData,
+    },
+  ).then((response) => {
+    if (response.ok) {
+      showMessage(successMessageShow);
+      restoreParameters();
+    } else {
+      showMessage(errorMessageShow);
+    }
+  }).catch(() => {
+    showMessage(errorMessageShow);
+  });
+});
